+++ lib/owoify_dart.dart
@@ -1,8 +1,5 @@
-/// Support for doing something awesome.
-///
-/// More dartdocs go here.
+/// owoify_dart is a Dart port of [mohan-cao's owoify-js](https://github.com/mohan-cao/owoify-js) library.
+/// This library will turn your input string (English only) into nonsensical babyspeaks.
 library owoify_dart;
 
 export 'src/owoify_dart_base.dart';
-
-// TODO: Export any libraries intended for clients of this package.