+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.1.0
+**This version changed some mappings in presets, so the same input probably won't return the same result as version 2.0.0.**
+- Add more word mappings, so it's on par with mohan-cao's latest owoify.
+- Add uwuify and uvuify for convenience.
+
 ## 2.0.1
 - Minor fix on README.md to specify the correct version.